refactor: deduplicate filterMetadata, fix empty-string edge case
Move applyMetadataFilter → filterMetadata in core/src/notes.ts
(shared by MCP and REST). Empty include_metadata= query param now
returns all metadata (was silently stripping all).
